refactor: rename persist -> snapshot
