+++ src/services/statefuls/cache/manager.rs
@@ -6,9 +6,9 @@ use crate::services::cluster::replications::replication::ReplicationInfo;
 use crate::services::query_io::QueryIO;
 use crate::services::statefuls::cache::ttl::actor::TtlActor;
 use crate::services::statefuls::cache::ttl::manager::TtlSchedulerInbox;
-use crate::services::statefuls::persist::endec::encoder::encoding_processor::SaveActor;
-use crate::services::statefuls::persist::endec::encoder::encoding_processor::SaveTarget;
-use crate::services::statefuls::persist::DumpFile;
+use crate::services::statefuls::snapshot::endec::encoder::encoding_processor::SaveActor;
+use crate::services::statefuls::snapshot::endec::encoder::encoding_processor::SaveTarget;
+use crate::services::statefuls::snapshot::DumpFile;
 use anyhow::Result;
 use std::time::SystemTime;
 use std::{hash::Hasher, iter::Zip};
