+++ src/services/statefuls/mod.rs
@@ -1,2 +1,2 @@
 pub mod cache;
-pub mod persist;
+pub mod snapshot;
