+++ src/services/statefuls/snapshot/dump_loader.rs
