+++ src/services/statefuls/snapshot/encoding_command.rs
