+++ src/services/statefuls/snapshot/endec/encoder/mod.rs
